requirements.txt
# Core dependencies
python>=3.10
asyncio-mqtt>=0.13.0
aiohttp>=3.9.0
aioredis>=2.0.0

# Database
taospy>=3.0.0
taos-ws-py>=0.3.0

# Exchange API
ccxt>=4.0.0

# AI & Agent Framework
langgraph>=0.0.50
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
openai>=1.10.0
anthropic>=0.18.0

# Charting & Visualization
plotly>=5.18.0
matplotlib>=3.8.0
pillow>=10.0.0

# Message Queue
redis>=5.0.0

# Configuration & Environment
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Logging & Monitoring
loguru>=0.7.0
structlog>=24.1.0

# Security
cryptography>=41.0.0
python-jose>=3.3.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.12.0
ruff>=0.1.0
mypy>=1.8.0

# Utilities
numpy>=1.26.0
pandas>=2.1.0
tenacity>=8.2.0
